[TASK] Product policy

+++ ruppy3s2/app/Http/Controllers/ProductController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use Illuminate\Http\Request;
 use App\Models\Category;
 use App\Models\Product;
+use Illuminate\Support\Facades\Auth;
 
 class ProductController extends Controller
 {
@@ -26,12 +27,18 @@ class ProductController extends Controller
 
     public function store(Request $request)
     {
-        Product::create($request->all());
+        $input = $request->all();
+        $input['user_id'] = Auth::id();
+        Product::create($input);
         return redirect(route('products.index'));
     }
 
     public function edit(Product $product)
     {
+        if (Auth::user()->cannot('update', $product)) {
+            abort(403);
+        }
+
         $categories = Category::get();
         return view('products.edit', [
             'product' => $product,
@@ -41,12 +48,18 @@ class ProductController extends Controller
 
     public function update(Request $request, Product $product)
     {
+        if (Auth::user()->cannot('update', $product)) {
+            abort(403);
+        }
         $product->update($request->all());
         return redirect(route('products.index'));
     }
 
     public function destroy(Product $product)
     {
+        if (Auth::user()->cannot('delete', $product)) {
+            abort(403);
+        }
         $product->delete();
         return redirect(route('products.index'));
     }
